Opening bid. Lots more to do. Like, A LOT.
